babzel-tools/src/main/java/org/babzel/tools/opennlp/conllu/convert/ConlluPOSSamplesConverter.java
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.babzel.tools.opennlp.conllu.convert;

import io.vavr.Tuple;
import io.vavr.Tuple2;
import io.vavr.collection.Seq;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;
import opennlp.tools.postag.POSSample;
import org.babzel.tools.opennlp.conllu.parser.ConlluSentence;
import org.babzel.tools.opennlp.conllu.parser.ConlluWordLine;
import org.babzel.tools.opennlp.conllu.util.ConlluNormalizer;
import org.babzel.tools.opennlp.conllu.util.ConlluValidator;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class ConlluPOSSamplesConverter implements ConlluSamplesConverter<POSSample> {
    @NonNull
    private final ConlluNormalizer normalizer;
    @NonNull
    private final ConlluValidator validator;

    @Override
    public Seq<POSSample> convert(@NonNull Seq<ConlluSentence> sentences, @NonNull String language) {
        return sentences
                .map(sentence -> normalizer.normalizeSentence(sentence, language))
                .filter(validator::isValidForLemmatization)
                .map(ConlluSentence::flattenWords)
                .map(this::convert);
    }

    private POSSample convert(ConlluSentence sentence) {
        var posTags = getPosTags(sentence.getWords());
        return new POSSample(
                posTags.map(Tuple2::_1).toJavaList(),
                posTags.map(Tuple2::_2).toJavaList());
    }

    private Seq<Tuple2<String, String>> getPosTags(Seq<ConlluWordLine> words) {
        return words.map(this::getPosTag);
    }

    private Tuple2<String, String> getPosTag(ConlluWordLine word) {
        return word.isCompound()
                ? Tuple.of(word.joinSubForms(), word.joinPosTags())
                : Tuple.of(word.getForm(), word.getPosTag());
    }
}
